Refactor/Major: move web and sock servers into modules
